Update README.md
This is a Netflix Clone built using HTML and CSS, designed to replicate the visual aesthetics of Netflix’s homepage. It includes a responsive layout, a hero section with a background image, a call-to-action (CTA) button, and multiple content sections showcasing movies and TV shows. This project is purely front-end, focusing on UI/UX design without backend functionality.